Transferred theme store to the layout
